helper/schema: FieldReader needs to return computed status
